Implement a proper skb_pull for the future.
We never send two messages in a row currently, there's always a strict
ping-pong. This flushes the socket and thus skb_pull is never used
to extract the current message, leaving the next one in the pipe.
However, we do not want to forget about the skb_buff being fake and
be frustrated in the future.